feat: argument resolver 수정

+++ tle/tle-application/src/main/kotlin/com/example/tleapplication/support/security/UserArgumentResolver.kt
@@ -32,7 +32,8 @@ class UserArgumentResolver(
         val claims: Claims =
             claimExtractor.extractAllClaims(token) ?: throw ClaimNotFoundException()
 
-        val userId = claims["id"] as Long
+        val userId = (claims["id"] as Double).toLong()
+
         return AuthInfo(userId)
     }
 }
